requirements.txt
python-dotenv
OpenAI
aiogram